Compiler
Another go at solving the compiler problem. I've landed on the idea of basically copying the way the tokenizer does things. Have the compilation engine run through each line figuring out with token and type it is then copying over the xml along with the new tags required. Might be overkill doing it twice, but get it to work first.